requirements.txt
Django~=4.1.5
djangorestframework~=3.14.0
django-filter~=22.1
cryptography~=3.4.8
pip~=23.0
attrs~=22.2.0
wheel~=0.37.1
Pillow~=9.0.1
Jinja2~=3.1.2
keyring~=23.5.0
pytz~=2022.1
asgiref~=3.6.0
sqlparse~=0.4.3
colorama~=0.4.4
PyYAML~=5.4.1
requests~=2.25.1
djoser~=2.1.0
MarkupSafe~=2.1.1
httplib2~=0.20.2
certifi~=2022.12.7
itypes~=1.2.0
coreapi~=2.3.3
coreschema~=0.0.4
click~=8.0.3
uritemplate~=4.1.1
six~=1.16.0
packaging~=23.0
inflection~=0.5.1
oauthlib~=3.2.0
blinker~=1.4
idna~=3.3
urllib3~=1.26.5
chardet~=4.0.0
setuptools~=59.6.0
future~=0.18.2
PyJWT~=2.3.0
bcrypt~=3.2.0
django-debug-toolbar~=3.8.1
Pygments~=2.14.0